Stripe payment success added.

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { AfterViewInit, Component, ElementRef, Input, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { NavigationExtras, Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
@@ -7,19 +7,62 @@ import { IBasket } from 'src/app/shared/models/basket';
 import { IOrder } from 'src/app/shared/models/order';
 import { CheckoutService } from '../checkout.service';
 
+declare var Stripe;
+
 @Component({
   selector: 'app-checkout-payment',
   templateUrl: './checkout-payment.component.html',
   styleUrls: ['./checkout-payment.component.scss']
 })
-export class CheckoutPaymentComponent implements OnInit {
+export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
 
   @Input() checkoutForm: FormGroup;
+  @ViewChild('cardNumber', {static: true}) cardNumberElement: ElementRef;
+  @ViewChild('cardExpiry', {static: true}) cardExpiryElement: ElementRef;
+  @ViewChild('cardCvc', {static: true}) cardCvcElement: ElementRef;
+
+  stripe: any;
+  cardNumber: any;
+  cardExpiry: any;
+  cardCvc: any;
+  cardErrors: any;
+  cardHandler = this.onChange.bind(this); // handler to show error
 
   constructor(private basketService: BasketService, private checkoutService: CheckoutService,
               private toastr: ToastrService, private router: Router) { }
 
-  ngOnInit(): void {
+
+  ngAfterViewInit(): void {
+    this.stripe = Stripe('pk_test_51IMrBOJzRQfscld3IAZG36YKXqV60p9pSjfaMpjFGDBjUqfbPHs8c30pqPGt3Up43JsYe9DuNn0Pyg16grjZ9aFF00swMbM6sp');
+    const elements = this.stripe.elements();
+
+    this.cardNumber = elements.create('cardNumber'); // stripe element functionality
+    this.cardNumber.mount(this.cardNumberElement.nativeElement);
+    this.cardNumber.addEventListener('change', this.cardHandler);
+
+    this.cardExpiry = elements.create('cardExpiry'); // stripe element functionality
+    this.cardExpiry.mount(this.cardExpiryElement.nativeElement);
+    this.cardExpiry.addEventListener('change', this.cardHandler);
+
+    this.cardCvc = elements.create('cardCvc'); // stripe element functionality
+    this.cardCvc.mount(this.cardCvcElement.nativeElement);
+    this.cardCvc.addEventListener('change', this.cardHandler);
+  }
+
+  ngOnDestroy(): void {
+    this.cardNumber.destroy();
+    this.cardExpiry.destroy();
+    this.cardCvc.destroy();
+  }
+
+  // syntax here is called destructuring.
+  onChange({error}){
+  if (error){
+    this.cardErrors = error.message;
+  } else
+  {
+    this.cardErrors = null;
+  }
   }
 
   submitOrder(){
@@ -27,9 +70,25 @@ export class CheckoutPaymentComponent implements OnInit {
     const orderToCreate = this.OrderToCreate(basket);
     this.checkoutService.createOrder(orderToCreate).subscribe((order: IOrder) => {
       this.toastr.success('Order created successfully');
-      this.basketService.deleteLocalBasket(basket.id);
-      const navigationExtras: NavigationExtras = {state: order};
-      this.router.navigate(['checkout/success'], navigationExtras);
+      this.stripe.confirmCardPayment(basket.clientSecret, {
+        payment_method: {
+          card: this.cardNumber,
+          billing_details: {
+            name: this.checkoutForm.get('paymentForm').get('nameOnCard').value
+          }
+        }
+      }).then(result => {
+        console.log(result);
+        if (result.paymentIntent){
+            this.basketService.deleteLocalBasket(basket.id);
+            const navigationExtras: NavigationExtras = {state: order};
+            this.router.navigate(['checkout/success'], navigationExtras);
+        } else {
+          this.toastr.error(result.error.message);
+        }
+
+      });
+
     }, error => {
       this.toastr.error(error.message);
       console.log(error);
